+++ src/Match.tsx
@@ -1,43 +1,31 @@
-import React from "react";
+import React, { useContext } from "react";
 import Route from "./Route";
 import { RouteProps } from "./RouteProps";
 import { RouterContext, RouterContextType } from "./RouterContext";
-import { RouterException } from "./RouterException";
+import { useRouterContext } from "./useRouterContext";
 import { calculateMatch, sanitizeChildren } from "./utils";
 
 /**
  * Component for matching a path and passing the result to it's children.
  */
-export default class Match extends React.Component<RouteProps> {
-	public static propTypes: object;
+export default function Match(props: RouteProps) {
+	const context = useRouterContext("Match");
+	const location = props.location || context.location;
+	const match = calculateMatch(props, context);
 
-	public render(): JSX.Element {
-		return (
-			<RouterContext.Consumer>
-				{(context) => {
-					if (!context) {
-						throw RouterException("Match");
-					}
-					const location = this.props.location || context.location;
-					const match = calculateMatch(this.props, context);
+	const contextProps: RouterContextType = { ...context, location, match };
 
-					const props: RouterContextType = { ...context, location, match };
+	let { children } = props;
+	children = sanitizeChildren("Match", children, contextProps, props.path);
 
-					let { children } = this.props;
-					children = sanitizeChildren("Match", children, props, this.props.path);
-
-					if (!children) {
-						return null;
-					}
-					return (
-						<RouterContext.Provider value={props}>
-							{children}
-						</RouterContext.Provider>
-					);
-				}}
-			</RouterContext.Consumer>
-		);
+	if (!children) {
+		return null;
 	}
+	return (
+		<RouterContext.Provider value={contextProps}>
+			{children}
+		</RouterContext.Provider>
+	);
 }
 
 if (process.env.NODE_ENV !== "production") {